Added objects for MartConfiguration System.
Engine now has getMartConfiguration method that returns
a MartConfiguration object. This is a set of Datasets with the
starBases and primaryKeys set, but no Attribute/Filter groupings.

fixed MartExplorerTool to use the MartConfiguration, and,
until we get the fully functional system going, to accept
tableConstraint.fieldName in attribute and field requests.