Note, will need to add jdom to the distribution.
config.MartDTDEntityResolver -- impliments the org.xml.sax.EntityResolver
method.  Overides the resolveEntity method to specifically
handle the DOCTYPE declaration for the mart dtd.  In this case,
it returns an org.xml.sax.InputSource object created from a byteStream
obtained from a database call.  For other xml entities, the method
returns null, so that the parser can invoke its default handler.

config.MartXMLutils -- utility class with static methods to get InputSources
for the mart xml and dtd documents stored in the mart database, and
store xml and dtd files into the database.

config.XMLLoader -- commandline tool to load xml and dtd files from the system into the
mart database.

test.Base -- added method to get a database connection using the testconnection.conf properties.

test.ConfigurationTest -- will contain tests for all aspects of the configuration system. Currently tests
the round trip of xml files out of and into the database.
First it retrieves an xml and parses it with jdom, then stores it back into the
database.  It then retrieves it again, and compares it to the version it originally retrieved.
This tests the xml validity, along with the functionality of the MartXMLutils static methods
getInputSourceFor() and storeConfiguration().

debugged the MartConfigurationTemplate.xml and dtd