placed context sensitive completion choice Sets into a TreeMap,
created public method allowing client programs to add Collections
of items to specific choice Sets in the TreeMap, mapped by a
String key.  This puts the responsibility for setting up the
context of the completion system on the client, rather than
the MartCompleter object  MartCompleter now only adds
items from the MartConfiguration to the choice sets.