MultiDSViewAdaptor is an interface that extends DSViewAdaptor by adding a boolean removeDatasetView(DatasetView) method to the interface.  This interface is implimented by any DSViewAdaptor object that will hold more than one DatasetView
DatabaseDSViewAdaptor now impliments MultiDSViewAdaptor,
and getMartLocations method.  Also has a setDatabasePassword(String)
method that allows client code to set the Password for the DataSource
used to create the DatabaseDSViewAdaptor, so that when
getMartLocations is called, the MartLocation element contains
the password.  Otherwise, the resulting MartLocation element
will not contain a password, even if one was used to connect
to the DataSource.  Documentation explains the potential security
risk of using passwords in this way.