Can now query against Oracle mart.
implimented JDBC abstracted methods in place of
mySQL specific commands such as 'describe table', etc.

Many comparisons between values in the XML and
values as they are presented from the RDBMS (eg,
tablenames and field names) needed to be 'normalized' to
always compare toLowerCase() versions of the strings.  The
system always carries these values as they appear in the DB.

Also solved the annoying printout from the Ewin connection pooling
software.

This only works with the Oracle Thin driver, as it is most like
the other database drivers in terms of connection URL syntax.

Needs further testing