works with new DatasetConfigIterator passed by getDatasetConfigs() from DSConfigAdaptor.
refactored MartShellLib to act as controller for both MartShell
and MartCompleter (eg, envMart, envDataset, localDataset, etc.
are all in MartShellLib).  This makes it easier to track when and
where DatasetConfig objects are stored in memory within Shell.
There still may be some tweaking that can be done to reduce
the number of DatasetConfig objects that MartShell holds in memory
during a single query cycle.