moved DatabaseDialog to org.ensembl.mart.guiutils
Modified to ignore hidden elements when parsing XML, using
the new Adaptor system

DatabaseConnection dialog interaction now requires that
a user enter correct information, or cancel, before being
allowed to proceed with the use of a connection. If not,
they are requested to try again, and the dialog is launched
again. I also added a ConnectionName field, allowing users
to set the name of the Connection in their drop down list of
choices.  This is consistent with MartShell.