moved DatabaseDialog to org.ensembl.mart.guiutils
All Adaptors created are instructed to completely ignore cache, and include hidden
elements.

Database connection dialog interaction requires that the
user enter correct information or cancel in order to proceed
with a connection.  If they do not enter correct connection
information, they are instructed to try again, and presented
with the dialog again.