removed SequenceTest and replaced it with Sequencetst. This prevents it from running in the ant unit tests until we fix it.
Base now contains the single DetailedDataSource that all
test cases should utilize for database connections. This
is specified only in the connection.properties file, with no
default behavior.  setUp actually fails a test if this file doesnt
exist.  Base also contains a connected method that it and child
classes can use to assert that a DetailedDataSource is actually
connected to a database.

DetailedDataSource now extends Base and does very
simple tests on the DetailedDataSource constructor,
using connection parameters specified in the connection.properties
file parsed in Base. Commented out a test that was broken,
but not documented well enough to fix it.