countFocus for a Query with no filters now executes   select (count distinct pkeys[0]) from mainTables[0];
unless no DataSource, primaryKeys, and mainTables
have been set in the query (eg, a mart and dataset have
not been chosen).

countRows for a Query with no attributes or filters executes
  select count(*) from mainTables[0];

unless no DataSource, primaryKeys, and mainTables have
been set in the query (eg, a mart and dataset have not
been chosen).

also contains a bugfix