added busy icon to methods that get or manipulate RDBMS connections.
Also fixed a bug when a nieve user hits the update button
when there are no datasetconfigs imported and open. Previously
it printed a stack trace for a null pointer exception but didnt
really inform the user that this is not correct. Now informs
the user to import a datasetconfig first.