first phase of templating implementation: (a) Export on export the meta_template__template__main and meta_template__xml__dm tables are created if they do not already exist.
The template setting for the dataset_id_key is then
stored in the main table.

If there are 0 or 1 datasets using this template setting
then a new template is created from the present config
with certain attributes such as field and tableConstraint
blanked and stored in the xml dm table

If there is more than 1 dataset using it a method to be
implemented will be called which will merge the dataset
and template appropiately before storing both the new
datasetconfig and template.

(b) Import template

Template XML can be opened for editing

(c) Export template

Stores the opened template XML. Needs a method to update
all the datasetConfigs using the template