MARTBUILDER UPDATE - WORK IN PROGRESS. Initial steps towards rewrite of internals for 0.6 release, allowing for the creation of new features required for Chado. Several chunks of functionality have been removed or are commented, but will be reinstated one-by-one over the next few days. These are: restricted relations, restricted tables, concat relations, expression columns, partitioned columns, partitioned schemas, and inherit-override columns. New features in this commit are: compound relations, merged dimensions, masked dimensions, forced relations.
NOTE: Due to changes in the internals, 0.5 XML will load but any user customisations to datasets will probably not persist unless saved in the 0.6  XML format, which is NOT FINALIZED YET so don't rely on it! In general the 0.6 XML will be much simpler than 0.5, but it will not be possible to write a converter that can translate 100% of user dataset modifications, although it will do its best.